move new job to root loop
