+++ validator/control_node/src/synthetics/refresh_synthetic_data.py
@@ -1,12 +1,8 @@
 import asyncio
-import random
-import requests
-import os
 import datetime
 import json
 import nltk
 import time
-import fcntl
 from pydantic import BaseModel
 from core import task_config as tcfg
 from redis.asyncio import Redis
@@ -53,45 +49,7 @@ async def update_tasks_synthetic_data(redis_db: Redis, slow_sync: bool = True, f
             if slow_sync:
                 await asyncio.sleep(0.1)
 
-async def fetch_random_text():
-    n_paragraphes = random.randint(2, 4)
-    n_sentences = random.randint(1, 6)
-    url = f'http://metaphorpsum.com/paragraphs/{n_paragraphes}/{n_sentences}'
-    response = requests.get(url)
-    if response.status_code == 200:
-        return response.text, n_paragraphes, n_sentences
-    else:
-        raise logger.error(f"Failed to fetch text from metaphorpsum.com: {response.status_code}")
-
-async def get_save_random_text() -> None:
-    if not os.path.exists(scst.RANDOM_TEXT_FILE):
-        open(scst.RANDOM_TEXT_FILE, 'w').close()
-        
-    while True:
-        try:
-            with open(scst.RANDOM_TEXT_FILE, 'r') as file:
-                lines = file.readlines()
-            queue_size = len(lines)            
-            if queue_size < 500:
-                text, n_paragraphes, n_sentences = await fetch_random_text()
-                n_words = len(text.split())                
-                with open(scst.RANDOM_TEXT_FILE, 'a') as file:
-                    try:
-                        fcntl.flock(file, fcntl.LOCK_EX)
-                        file.write(text + '\n')
-                        logger.debug(f"Pushed random metaphorpsum.com text with {n_words} words, {n_paragraphes} paragraphs, and {n_sentences} sentences to text file")
-                    finally:
-                        fcntl.flock(file, fcntl.LOCK_UN)
-            else:
-                logger.debug(f"Text file '{scst.RANDOM_TEXT_FILE}' is full. Skipping text insertion")                
-            await asyncio.sleep(1)
-            
-        except Exception as e:
-            await asyncio.sleep(60)
-            logger.error(f"Error fetching and saving synthetic data: {e} - sleeping for 60s")
-
 async def continuously_fetch_synthetic_data_for_tasks(redis_db: Redis) -> None:
-    asyncio.create_task(get_save_random_text())
     await update_tasks_synthetic_data(redis_db, slow_sync=False)
     while True:
         await update_tasks_synthetic_data(redis_db, slow_sync=True)
